requirements.txt
python-binance==1.0.17
PyQt6==6.5.2
python-dotenv==1.0.0
aiohttp==3.8.5